# Back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Matheus Henrique",
        "description": "Documentação gerada automaticamente com L5-Swagger.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "description": "Endpoint for user authentication. Requires registration and password.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "registration",
                                    "password"
                                ],
                                "properties": {
                                    "registration": {
                                        "type": "string",
                                        "format": "registration",
                                        "example": "user123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully"
                    },
                    "400": {
                        "description": "registration was not provided | The user's registration must be of the type String | password was not provided"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout do usuário autenticado",
                "description": "Encerra a sessão do usuário autenticado removendo o token de acesso atual (Sanctum).",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Logged out"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nenhum token encontrado para exclusão",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "No token found"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Not authenticated"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}